任务优化

+++ 51fbapi-biz/src/main/java/com/ald/fanbei/api/biz/service/AfTaskBrowseGoodsService.java
@@ -17,5 +17,5 @@ public interface AfTaskBrowseGoodsService{
 
     int addBrowseGoodsTaskUser(Long userId, Long goodsId);
 
-    Integer addBrowseGoodsTaskUserRecord(Long userId, Long goodsId);
+    Long addBrowseGoodsTaskUserRecord(Long userId, Long goodsId);
 }
